backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
async-timeout==4.0.3
beautifulsoup4==4.13.4
cachetools==6.1.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
distro==1.9.0
exceptiongroup==1.3.0
extruct==0.18.0
fastapi==0.116.1
frozendict==2.4.6
greenlet==3.2.3
h11==0.16.0
html5lib==1.1
html_text==0.7.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
isodate==0.7.2
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jstyleson==0.0.2
langchain==0.3.26
langchain-core==0.3.71
langchain-text-splitters==0.3.8
langsmith==0.4.8
lxml==6.0.0
lxml_html_clean==0.4.2
mf2py==2.0.1
openai==1.97.1
orjson==3.11.0
packaging==25.0
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
PyLD==2.0.4
pyparsing==3.2.3
pyRdfa3==3.6.4
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
rdflib==7.1.4
recipe_scrapers==15.8.0
requests==2.32.4
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
starlette==0.47.2
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
w3lib==2.3.1
webencodings==0.5.1
zstandard==0.23.0
